# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EV Database API
  description: |
    A comprehensive API for managing electric vehicle data, contributions, and user interactions.

    ## Authentication

    This API supports two authentication methods:

    ### 1. Frontend Authentication (JWT)
    - Used by the frontend application
    - Requires `X-Frontend-Secret` header with the frontend secret
    - Uses JWT tokens for user authentication
    - No rate limiting applied

    ### 2. External API Authentication (API Key)
    - Used by external applications and integrations
    - Requires a valid API key in the `Authorization` header as `Bearer <api-key>`
    - Subject to rate limiting based on user role
    - Consumes API credits per request

    ## Rate Limiting

    API key requests are subject to rate limiting:
    - **MEMBER**: 100 requests per hour
    - **MODERATOR**: 500 requests per hour
    - **ADMIN**: 1000 requests per hour

    ## Error Responses

    The API uses standard HTTP status codes and returns error responses in the following format:
    ```json
    {
      "error": "Error message description"
    }
    ```

  version: 1.0.0
  contact:
    name: EV Database API Support
    email: support@evdatabase.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.evdatabase.com/api
    description: Production server
security:
  - FrontendAuth: []
  - ApiKeyAuth: []
paths:
...
  /vehicles:
    get:
      summary: Get All Vehicles
      description: Get a paginated list of all vehicles with optional filtering
      tags: [Vehicles]
      security: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search term for make, model, or description
          schema:
            type: string
        - name: make
          in: query
          description: Filter by vehicle make
          schema:
            type: string
        - name: year
          in: query
          description: Filter by vehicle year
          schema:
            type: integer
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [id, make, model, year, batteryCapacity, range]
            default: id
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create New Vehicle
      description: Create a new vehicle (admin only)
      tags: [Vehicles]
      security:
        - FrontendAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '201':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    FrontendAuth:
      type: apiKey
      in: header
      name: X-Frontend-Secret
      description: Frontend secret for internal application use

    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key for external access
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          description: Unique vehicle identifier
        make:
          type: string
          description: Vehicle manufacturer
        model:
          type: string
          description: Vehicle model
        year:
          type: integer
          description: Vehicle year
        batteryCapacity:
          type: integer
          nullable: true
          description: Battery capacity in kWh
        range:
          type: integer
          nullable: true
          description: Vehicle range in km
        chargingSpeed:
          type: integer
          nullable: true
          description: Charging speed in kW
        description:
          type: string
          nullable: true
          description: Vehicle description
      required: [id, make, model, year]
    VehicleInput:
      type: object
      description: Input schema for creating/updating vehicles
      properties:
        make:
          type: string
          description: Vehicle manufacturer
        model:
          type: string
          description: Vehicle model
        year:
          type: integer
          minimum: 1900
          maximum: 2030
          description: Vehicle year
        batteryCapacity:
          type: integer
          nullable: true
          minimum: 1
          description: Battery capacity in kWh
        range:
          type: integer
          nullable: true
          minimum: 1
          description: Vehicle range in km
        chargingSpeed:
          type: integer
          nullable: true
          minimum: 1
          description: Charging speed in kW
        description:
          type: string
          nullable: true
          description: Vehicle description
      required: [make, model, year]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page
      required: [page, limit, total, totalPages, hasNext, hasPrev]
...
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Authentication required"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Admin access required"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Validation failed"
...